change prints to logger+ some improvements

+++ meg_data_processing/03-processing_anatomical_data.py
@@ -8,11 +8,20 @@ from typing import List
 
 import mne
 import nibabel as nib
-from mne.parallel import parallel_func
 
 sys.path.append('../../MFRS/')
 from utils.config import study_path, subjects_dir, spacing
 from utils.arg_parser import source_rescontruction_parser
+
+# Import the logging module and configure it
+import logging
+
+logging.basicConfig(
+    level=logging.INFO,
+    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
+)
+logger = logging.getLogger(__name__)
+
 
 def tee_output(command: List[str], log_file: str) -> None:
     """
@@ -33,8 +42,8 @@ def tee_output(command: List[str], log_file: str) -> None:
     """
     # Join the command elements to create a readable command string
     to_print = " ".join(command)
-    print("Running:\n")
-    print(to_print)
+    logger.info("Running:\n")
+    logger.info(to_print)
 
     # Open the log_file for writing the output of the command
     with open(log_file, 'wb') as fid:
@@ -66,7 +75,7 @@ def process_subject_anat(subject_id: int) -> None:
     """
 
     subject = f"sub-{subject_id:02d}"
-    print(f"Processing {subject}")
+    logger.info(f"Processing {subject}")
 
     t1_fname = os.path.join(study_path, 'ds000117', subject, 'ses-mri/anat',
                        f'{subject}_ses-mri_acq-mprage_T1w.nii.gz')
@@ -74,14 +83,14 @@ def process_subject_anat(subject_id: int) -> None:
     subject_dir = os.path.join(subjects_dir, subject)
 
     if os.path.isdir(subject_dir):
-        print('  Skipping reconstruction (folder exists)')
+        logger.info('Skipping reconstruction (folder exists)')
     else:
-        print('  Running reconstruction (usually takes hours)')
+        logger.info('Running reconstruction (usually takes hours)')
         t0 = time.time()
         tee_output(
             ['recon-all', '-all', '-s', subject, '-sd', subjects_dir,
              '-i', t1_fname], log_fname)
-        print(f'  Recon for {subject} complete in {((time.time() - t0) / 60. / 60.):0.1f} hours')
+        logger.info(f'Recon for {subject} complete in {((time.time() - t0) / 60. / 60.):0.1f} hours')
 
     bem_dir = os.path.join(subjects_dir, subject, 'bem')
 
@@ -94,29 +103,29 @@ def process_subject_anat(subject_id: int) -> None:
     # performed by MNE python functions mne.bem.make_watershed_bem
     if not (os.path.isfile(sbj_inner_skull_fname) or
             os.path.isfile(inner_skull_fname)):
-        print(f"{inner_skull_fname} ---> FILE NOT FOUND!!!---> BEM computed")
+        logger.info(f"{inner_skull_fname} ---> FILE NOT FOUND!!!---> BEM computed")
         mne.bem.make_watershed_bem(subject, subjects_dir, overwrite=True, verbose=True)
     else:
-        print(f"\n*** inner skull {inner_skull_fname} surface exists!!!\n")
+        logger.info(f"\n inner skull {inner_skull_fname} surface exists!!!\n")
 
     # Create a BEM model for a subject        
     fname_bem_surfaces = os.path.join(bem_dir, f'{subject}-5120-bem.fif')
     if not os.path.isfile(fname_bem_surfaces):
-        print(f'  Setting up 1-layer BEM')
+        logger.info(f'Setting up 1-layer BEM')
         conductivity = (0.3, 0.006, 0.3)[:1]
         try:
             bem_surfaces = mne.make_bem_model(
                 subject, ico=4, conductivity=conductivity,
                 subjects_dir=subjects_dir)
         except RuntimeError as exp:
-            print(f'  FAILED to create 1-layer BEM for {subject}: {exp.args[0]}')
+            logger.info(f'FAILED to create 1-layer BEM for {subject}: {exp.args[0]}')
         # Write BEM surfaces to a fiff file
         mne.write_bem_surfaces(fname_bem_surfaces, bem_surfaces)
 
     # Create a BEM solution using the linear collocation approach
     fname_bem = os.path.join(bem_dir, f'{subject}-5120-bem-sol.fif')
     if not os.path.isfile(fname_bem):
-        print(f'  Computing  1-layer BEM solution')
+        logger.info('Computing 1-layer BEM solution')
         bem_model = mne.read_bem_surfaces(fname_bem_surfaces)
         bem = mne.make_bem_solution(bem_model)
         mne.write_bem_solution(fname_bem, bem)
@@ -124,7 +133,7 @@ def process_subject_anat(subject_id: int) -> None:
     # Create the surface source space
     fname_src = os.path.join(subjects_dir, subject, 'bem', f'{subject}-{spacing}-src.fif')
     if not os.path.isfile(fname_src):
-        print('  Setting up source space')
+        logger.info('Setting up source space')
         src = mne.setup_source_space(subject, spacing, subjects_dir=subjects_dir)
         mne.write_source_spaces(fname_src, src)
 
@@ -132,24 +141,29 @@ if __name__=="__main__":
     parser = source_rescontruction_parser()
     args = parser.parse_args()
     
-    parallel, run_func, _ = parallel_func(process_subject_anat(args.subject), n_jobs=-1)
+    # Run the process_subject_anat function in parallel for all subjects
+    subject_id = args.subject
+    process_subject_anat(subject_id)
  
-    # now we do something special for fsaverage
+    # Now we do something special for fsaverage
     fsaverage_src_dir = os.path.join(os.environ['FREESURFER_HOME'], 'subjects', 'fsaverage')
     fsaverage_dst_dir = os.path.join(subjects_dir, 'fsaverage')
 
-    print('Copying fsaverage into subjects directory')  # to allow writting in folder
-    os.unlink(fsaverage_dst_dir)  # remove symlink
-    shutil.copytree(fsaverage_src_dir, fsaverage_dst_dir)
+    if os.path.exists(fsaverage_dst_dir):
+        logger.info('Copying fsaverage into subjects directory')  # to allow writing in folder
+        os.unlink(fsaverage_dst_dir)  # remove symlink
+        shutil.copytree(fsaverage_src_dir, fsaverage_dst_dir)
 
-    fsaverage_bem = os.path.join(fsaverage_dst_dir, 'bem')
-    if not os.path.isdir(fsaverage_bem):
-        os.mkdir(fsaverage_bem)
+        fsaverage_bem = os.path.join(fsaverage_dst_dir, 'bem')
+        if not os.path.isdir(fsaverage_bem):
+            os.mkdir(fsaverage_bem)
 
-    fsaverage_src = os.path.join(fsaverage_bem, 'fsaverage-5-src.fif')
-    if not os.path.isfile(fsaverage_src):
-        print('Setting up source space for fsaverage')
-        src = mne.setup_source_space('fsaverage', 'ico5', subjects_dir=subjects_dir)
-        for s in src:
-            assert np.array_equal(s['vertno'], np.arange(10242))
-        mne.write_source_spaces(fsaverage_src, src)
+        fsaverage_src = os.path.join(fsaverage_bem, 'fsaverage-5-src.fif')
+        if not os.path.isfile(fsaverage_src):
+            logger.info('Setting up source space for fsaverage')
+            src = mne.setup_source_space('fsaverage', 'ico5', subjects_dir=subjects_dir)
+            for s in src:
+                assert np.array_equal(s['vertno'], np.arange(10242))
+            mne.write_source_spaces(fsaverage_src, src)
+    else:
+        logger.warning("fsaverage_dst_dir does not exist, skipping copying of fsaverage")
